Samp/requirements.txt
torch==2.6.0
torchaudio==2.6.0
transformers==4.57.1
packaging
jinja2
openai-whisper
jsonlines
pandas
validators
megfile
sty
editdistance
zhconv
matplotlib
scikit-image
librosa
accelerate
aiohttp
peft
onnxruntime-gpu==1.23.2
omegaconf==2.3.0
funasr
sox
six==1.17.0
hyperpyyaml
pillow
sentencepiece
easydict
fire
ujson
opencv-python
imagesize
cairosvg
rich
wget
gdown
datasets
huggingface_hub
loguru
tqdm
decord
blobfile
timm
sacrebleu==1.5.1
absl-py 
datasets
flash-attn==2.7.1.post1